requirements.txt
torch==2.8.0
torchvision==0.20.1
datasets==2.19.0
Pillow==11.0.0
